Some code refactoring with the eventual goal of getting zap to build against stock xulrunner versions: * Minimize modifications to Mozilla code. * Change zap code from using internal Mozilla APIs to external ones. * Build the zap apis as shared external libs for all build configurations (libxul/non-libxul).

+++ mozilla/zap/zmk/src/zapSpeexAudioProcessor.h
@@ -78,14 +78,14 @@ private:
   
   PRBool mAEC;
   PRBool mAEC2Stage;
-  float mAECTail;
+  double mAECTail;
   PRBool mDenoise;
   PRBool mAGC;
-  float mAGCLevel;
+  double mAGCLevel;
   PRBool mVAD;
   PRBool mDereverb;
-  float mDereverbLevel;
-  float mDereverbDecay;
+  double mDereverbLevel;
+  double mDereverbDecay;
   
   SpeexEchoState* mEchoState;
   SpeexPreprocessState* mPreprocessState;
